feat(watcher): cross-compilation Makefile + fix ARM Nlink cast
- Makefile: add watcher-build-all (linux/amd64, linux/arm64, linux/arm/v7)
- Makefile: add watcher-clean, update watcher-build with -ldflags="-s -w"
- Makefile: add WATCHER_BIN, WATCHER_ARCHS, RED variables
- hardlink: cast stat.Nlink to uint64 for ARM compat (uint32 on arm/arm64)

+++ watcher/internal/hardlink/hardlink.go
@@ -17,7 +17,7 @@ func Info(path string) (FileInfo, error) {
 		return FileInfo{}, err
 	}
 	return FileInfo{
-		Nlink:    stat.Nlink,
+		Nlink:    uint64(stat.Nlink),
 		Inode:    stat.Ino,
 		DeviceID: uint64(stat.Dev),
 	}, nil
